feat: interface status impl

+++ src/main/java/com/onetuks/ihub/service/interfaces/InterfaceStatusService.java
@@ -4,12 +4,15 @@ import com.onetuks.ihub.dto.interfaces.InterfaceStatusCreateRequest;
 import com.onetuks.ihub.dto.interfaces.InterfaceStatusUpdateRequest;
 import com.onetuks.ihub.entity.interfaces.InterfaceStatus;
 import com.onetuks.ihub.entity.project.Project;
+import com.onetuks.ihub.entity.user.User;
 import com.onetuks.ihub.mapper.InterfaceStatusMapper;
 import com.onetuks.ihub.repository.InterfaceStatusJpaRepository;
 import com.onetuks.ihub.repository.ProjectJpaRepository;
+import com.onetuks.ihub.service.project.ProjectMemberCheckComponent;
 import jakarta.persistence.EntityNotFoundException;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,15 +20,15 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class InterfaceStatusService {
 
+  private final ProjectMemberCheckComponent projectMemberCheckComponent;
   private final InterfaceStatusJpaRepository interfaceStatusJpaRepository;
   private final ProjectJpaRepository projectJpaRepository;
 
   @Transactional
-  public InterfaceStatus create(InterfaceStatusCreateRequest request) {
-    InterfaceStatus status = new InterfaceStatus();
-    InterfaceStatusMapper.applyCreate(status, request);
-    status.setProject(findProject(request.projectId()));
-    return interfaceStatusJpaRepository.save(status);
+  public InterfaceStatus create(User user, InterfaceStatusCreateRequest request) {
+    projectMemberCheckComponent.checkIsProjectOwner(user.getUserId(), request.projectId());
+    return interfaceStatusJpaRepository.save(
+        InterfaceStatusMapper.applyCreate(findProject(request.projectId()), request));
   }
 
   @Transactional(readOnly = true)
@@ -34,20 +37,23 @@ public class InterfaceStatusService {
   }
 
   @Transactional(readOnly = true)
-  public List<InterfaceStatus> getAll() {
-    return interfaceStatusJpaRepository.findAll();
+  public Page<InterfaceStatus> getAll(Pageable pageable) {
+    return interfaceStatusJpaRepository.findAll(pageable);
   }
 
   @Transactional
-  public InterfaceStatus update(String statusId, InterfaceStatusUpdateRequest request) {
+  public InterfaceStatus update(User user, String statusId, InterfaceStatusUpdateRequest request) {
     InterfaceStatus status = findEntity(statusId);
-    InterfaceStatusMapper.applyUpdate(status, request);
-    return status;
+    projectMemberCheckComponent.checkIsProjectOwner(
+        user.getUserId(), status.getProject().getProjectId());
+    return InterfaceStatusMapper.applyUpdate(status, request);
   }
 
   @Transactional
-  public void delete(String statusId) {
+  public void delete(User user, String statusId) {
     InterfaceStatus status = findEntity(statusId);
+    projectMemberCheckComponent.checkIsProjectOwner(
+        user.getUserId(), status.getProject().getProjectId());
     interfaceStatusJpaRepository.delete(status);
   }
